requirements.txt
anyio==3.6.2
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
fastapi==0.91.0
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.24.2
pandas==1.5.3
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.24.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
